+++ bots/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog for the Sesheta's Bot Cohort
 
+## [0.6.0] - 2018-Jul-04 - goern
+
+### Added
+
+The merger will now send out a notification to a mattermost channel.
+
+**merge_pr.py** will now send some metrics on merged PR to Prometheus, `pushgateway:9091` is used as a default value for the push gateway.
+
 ## [0.4.0] - 2018-May-03 - goern
 
 ### Added